Shib refactor: move builtin2shib to AdminIT #2939
