+++ src/main/java/edu/harvard/iq/dataverse/api/Admin.java
@@ -1,17 +1,25 @@
 package edu.harvard.iq.dataverse.api;
 
 import edu.harvard.iq.dataverse.Dataverse;
+import edu.harvard.iq.dataverse.EMailValidator;
+import edu.harvard.iq.dataverse.Shib;
 import edu.harvard.iq.dataverse.actionlogging.ActionLogRecord;
+import static edu.harvard.iq.dataverse.api.AbstractApiBean.errorResponse;
 import edu.harvard.iq.dataverse.api.dto.RoleDTO;
+import edu.harvard.iq.dataverse.authorization.AuthenticatedUserDisplayInfo;
 import edu.harvard.iq.dataverse.authorization.AuthenticatedUserLookup;
 import edu.harvard.iq.dataverse.authorization.AuthenticationProvider;
+import edu.harvard.iq.dataverse.authorization.UserIdentifier;
 import edu.harvard.iq.dataverse.authorization.exceptions.AuthenticationProviderFactoryNotFoundException;
 import edu.harvard.iq.dataverse.authorization.exceptions.AuthorizationSetupException;
 import edu.harvard.iq.dataverse.authorization.providers.AuthenticationProviderFactory;
 import edu.harvard.iq.dataverse.authorization.providers.AuthenticationProviderRow;
 import edu.harvard.iq.dataverse.authorization.providers.builtin.BuiltinAuthenticationProvider;
 import edu.harvard.iq.dataverse.authorization.providers.builtin.BuiltinUser;
+import edu.harvard.iq.dataverse.authorization.providers.builtin.BuiltinUserServiceBean;
 import edu.harvard.iq.dataverse.authorization.providers.shib.ShibAuthenticationProvider;
+import edu.harvard.iq.dataverse.authorization.providers.shib.ShibServiceBean;
+import edu.harvard.iq.dataverse.authorization.providers.shib.ShibUtil;
 import edu.harvard.iq.dataverse.authorization.users.AuthenticatedUser;
 import edu.harvard.iq.dataverse.engine.command.impl.PublishDataverseCommand;
 import edu.harvard.iq.dataverse.settings.Setting;
@@ -29,8 +37,10 @@ import javax.ws.rs.core.Response;
 import static edu.harvard.iq.dataverse.util.json.NullSafeJsonBuilder.jsonObjectBuilder;
 import static edu.harvard.iq.dataverse.util.json.JsonPrinter.*;
 import java.sql.SQLException;
+import java.util.Map;
 import java.util.logging.Level;
 import java.util.logging.Logger;
+import javax.ejb.EJB;
 import javax.ejb.Stateless;
 import javax.validation.ConstraintViolation;
 import javax.validation.ConstraintViolationException;
@@ -45,7 +55,12 @@ import javax.ws.rs.core.Response.Status;
 public class Admin extends AbstractApiBean {
     
     private static final Logger logger = Logger.getLogger(Admin.class.getName());
-    
+
+    @EJB
+    BuiltinUserServiceBean builtinUserService;
+    @EJB
+    ShibServiceBean shibService;
+
     @Path("settings")
     @GET
     public Response listAllSettings() {
@@ -343,6 +358,136 @@ public class Admin extends AbstractApiBean {
         output.add("email", builtinUser.getEmail());
         output.add("username", builtinUser.getUserName());
         return okResponse(output);
+    }
+
+    /**
+     * This is used in testing via AdminIT.java but we don't expect sysadmins to
+     * use this.
+     */
+    @Path("authenticatedUsers/convert/builtin2shib")
+    @PUT
+    public Response builtin2shib(String content) {
+        logger.info("entering builtin2shib...");
+        try {
+            AuthenticatedUser userToRunThisMethod = findAuthenticatedUserOrDie();
+            if (!userToRunThisMethod.isSuperuser()) {
+                return errorResponse(Response.Status.FORBIDDEN, "Superusers only.");
+            }
+        } catch (WrappedResponse ex) {
+            return errorResponse(Response.Status.FORBIDDEN, "Superusers only.");
+        }
+        boolean disabled = false;
+        if (disabled) {
+            return errorResponse(Response.Status.BAD_REQUEST, "API endpoint disabled.");
+        }
+        AuthenticatedUser builtInUserToConvert = null;
+        String emailToFind;
+        String password;
+        String authuserId = "0"; // could let people specify id on authuser table. probably better to let them tell us their 
+        try {
+            String[] args = content.split(":");
+            emailToFind = args[0];
+            password = args[1];
+//            authuserId = args[666];
+        } catch (ArrayIndexOutOfBoundsException ex) {
+            return errorResponse(Response.Status.BAD_REQUEST, ex.toString());
+        }
+        AuthenticatedUser existingAuthUserFoundByEmail = shibService.findAuthUserByEmail(emailToFind);
+        String existing = "NOT FOUND";
+        if (existingAuthUserFoundByEmail != null) {
+            builtInUserToConvert = existingAuthUserFoundByEmail;
+            existing = existingAuthUserFoundByEmail.getIdentifier();
+        } else {
+            long longToLookup = Long.parseLong(authuserId);
+            AuthenticatedUser specifiedUserToConvert = authSvc.findByID(longToLookup);
+            if (specifiedUserToConvert != null) {
+                builtInUserToConvert = specifiedUserToConvert;
+            } else {
+                return errorResponse(Response.Status.BAD_REQUEST, "No user to convert. We couldn't find a *single* existing user account based on " + emailToFind + " and no user was found using specified id " + longToLookup);
+            }
+        }
+        String shibProviderId = ShibAuthenticationProvider.PROVIDER_ID;
+        Map<String, String> randomUser = shibService.getRandomUser();
+//        String eppn = UUID.randomUUID().toString().substring(0, 8);
+        String eppn = randomUser.get("eppn");
+        String idPEntityId = randomUser.get("idp");
+        String notUsed = null;
+        String separator = "|";
+        UserIdentifier newUserIdentifierInLookupTable = new UserIdentifier(idPEntityId + separator + eppn, notUsed);
+        String overwriteFirstName = randomUser.get("firstName");
+        String overwriteLastName = randomUser.get("lastName");
+        String overwriteEmail = randomUser.get("email");
+        logger.info("overwriteEmail: " + overwriteEmail);
+        boolean validEmail = EMailValidator.isEmailValid(overwriteEmail, null);
+        if (!validEmail) {
+            // See https://github.com/IQSS/dataverse/issues/2998
+            return errorResponse(Response.Status.BAD_REQUEST, "invalid email: " + overwriteEmail);
+        }
+        /**
+         * @todo If affiliation is not null, put it in RoleAssigneeDisplayInfo
+         * constructor.
+         */
+        /**
+         * Here we are exercising (via an API test) shibService.getAffiliation
+         * with the TestShib IdP and a non-production DevShibAccountType.
+         */
+        idPEntityId = ShibUtil.testShibIdpEntityId;
+        String overwriteAffiliation = shibService.getAffiliation(idPEntityId, Shib.DevShibAccountType.RANDOM);
+        logger.info("overwriteAffiliation: " + overwriteAffiliation);
+        /**
+         * @todo Find a place to put "position" in the authenticateduser table:
+         * https://github.com/IQSS/dataverse/issues/1444#issuecomment-74134694
+         */
+        String overwritePosition = "staff;student";
+        AuthenticatedUserDisplayInfo displayInfo = new AuthenticatedUserDisplayInfo(overwriteFirstName, overwriteLastName, overwriteEmail, overwriteAffiliation, overwritePosition);
+        JsonObjectBuilder response = Json.createObjectBuilder();
+        JsonArrayBuilder problems = Json.createArrayBuilder();
+        if (password != null) {
+            response.add("password supplied", password);
+            boolean knowsExistingPassword = false;
+            BuiltinUser oldBuiltInUser = builtinUserService.findByUserName(builtInUserToConvert.getUserIdentifier());
+            if (oldBuiltInUser != null) {
+                String usernameOfBuiltinAccountToConvert = oldBuiltInUser.getUserName();
+                response.add("old username", usernameOfBuiltinAccountToConvert);
+                AuthenticatedUser authenticatedUser = shibService.canLogInAsBuiltinUser(usernameOfBuiltinAccountToConvert, password);
+                if (authenticatedUser != null) {
+                    knowsExistingPassword = true;
+                    AuthenticatedUser convertedUser = authSvc.convertBuiltInToShib(builtInUserToConvert, shibProviderId, newUserIdentifierInLookupTable);
+                    if (convertedUser != null) {
+                        /**
+                         * @todo Display name is not being overwritten. Logic
+                         * must be in Shib backing bean
+                         */
+                        AuthenticatedUser updatedInfoUser = authSvc.updateAuthenticatedUser(convertedUser, displayInfo);
+                        if (updatedInfoUser != null) {
+                            response.add("display name overwritten with", updatedInfoUser.getName());
+                        } else {
+                            problems.add("couldn't update display info");
+                        }
+                    } else {
+                        problems.add("unable to convert user");
+                    }
+                }
+            } else {
+                problems.add("couldn't find old username");
+            }
+            if (!knowsExistingPassword) {
+                problems.add("doesn't know password");
+            }
+//            response.add("knows existing password", knowsExistingPassword);
+        }
+
+        response.add("user to convert", builtInUserToConvert.getIdentifier());
+        response.add("existing user found by email (prompt to convert)", existing);
+        response.add("changing to this provider", shibProviderId);
+        response.add("value to overwrite old first name", overwriteFirstName);
+        response.add("value to overwrite old last name", overwriteLastName);
+        response.add("value to overwrite old email address", overwriteEmail);
+        if (overwriteAffiliation != null) {
+            response.add("affiliation", overwriteAffiliation);
+        }
+        response.add("problems", problems);
+        return okResponse(response);
     }
 
     @DELETE
